clean up retrofit dependency from API.kt file
